chore(#3038): account for turn bridge S1 split
